refactor some errors and test some feature(manual testing). add all DTOs for project(all entities). add @Valid validation for project(all entities). make All can register.

+++ src/main/java/com/example/producttestapi/service/CartServiceImpl.java
@@ -55,7 +55,7 @@ public class CartServiceImpl implements CartService {
                     Product product = cartItem.getProduct();
                     double currentProductPrice = product.getPrice();
                     if (cartItem.getPricePerItem() != currentProductPrice) {
-                        cartItem.setPricePerItem((int) currentProductPrice);
+                        cartItem.setPricePerItem( currentProductPrice);
                         cartItemRepo.save(cartItem);
                     }
                     return cartItem.getQuantityToTake() * currentProductPrice;
